Add task to push us to document our choices
Previously none of our config had any explanation of why it is there.
While it's possible to find some of this by digging through the commit
history, on the surface, what we have seems pretty arbitrary. This adds
a rake task that will fail the build of this repo unless all the config
has associated documentation, so that it can be audited in future.

This commit includes a (hopefully) non-contentious example of what such
documentation should look like. Basically: why???

This commit also disables all Rails cops. Although we want to include
the config, so it's validated as part of the build, leaving this checks
enabled leads to an invalid 'Rails/RakeEnvironment' issue.