Add documentation to all cops
This adds documentation to explain the origin of all the config we
have, which is now required by the 'explain_yourself' rake task.

Each item of config has a reference to its origin, which is either
the original GDS styleguide, or a particular commit. Although we
could use the git history to track down a particular item of config,
this has become quite hard in many cases, due to shifting config.

Many items of config have 'TODO' recommendations, where the config
is unnecessary (we're not in the business of specifying every single
piece of config RuboCop supports), or I've been unable to find any
explanation for its presence. Although it's not ideal to document
config as 'unclear', this still represents knowledge that required
effort to obtain, and will help someone auditing in future.