Replace Jenkins with GitHub Actions
This is largely based on the workflow for Jenkins [1]. Note that
that our release process is somewhat bespoke:

- We avoid pushing a gem version if it already exists
- We use the version in the gemspec to auto-create a tag

This makes it's difficult to reuse a 3rd party GitHub Action, which
we should avoid anyway due to the secret API token.

Other notes:

- The API token is automatically masked by GitHub in the log
output: "GEM_HOST_API_KEY: ***".

- We only need to trigger our tests on 'push' events, which
includes manual pushes (even before a PR is opened), as well
as merge commits to the master branch.

- We want to run tests (again) before releasing, since other
commits on the master branch may still cause an issue when a PR is
merged, even if there are no conflicts.

- Since the 'release' job only occurs on the master branch (outside of a
  PR), you need to look at the 'Actions' tab to debug it [2].

[1]: https://github.com/alphagov/govuk-jenkinslib/blob/master/vars/govuk.groovy
[2]: https://github.com/alphagov/rubocop-govuk/actions