Remove all config that matches the default
Previously we had a lot of config that did nothing to change the
defaults specified by RuboCop. While it may be interesting to know
that we are consistent with some of the defaults, these non-overrides
were incomplete, which is confusing, and took up a lot of space, which
makes it hard to see where we have actually deviated. This removes all
config that matches the RuboCop defaults, since it is redundant.