Partially enable Rails/Output
This still causes issues in some of our repos, where developers have
put output statements in DB migrations. Since a human is unlikely to
see the result of this output, we should go ahead and disallow this,
and manually resolve silence the issue for historical files, by using
a comment in the file itself (to avoid setting a precedent).