Disable Layout/FirstMethodArgumentLineBreak again
Previously we enabled this to try and bring consistency to repos
where auto-correct for other Cops was struggling due to the high
variation in layout we have in our repos. However, enabling this
Cop has lead to new layout issues with hanging brackets:

  Mailer.deliver(
      to: 'bob@example.com',
      from: 'us@example.com',
      subject: 'Important message',
      body: source.text)

This disables the Cop again, since enabling it contradicts the Ruby
styleguide [1], and has made layout harder to visually paerse in
some cases, which was not the original intention.

[1]: https://rubystyle.guide/#no-double-indent