Disable Layout/MultilineMethodArgumentLineBreaks
This cop is being disabled as it is one that deviates from the Rubocop
defaults and does not uniformly produce code that we can agree is
better.

The full backstory and discussion surrounding this removal is in
https://github.com/alphagov/rubocop-govuk/pull/63.

In a nutshell though, this cop offers a somewhat opinionated approach to
writing methods which conflicts with common idiomatic approaches such
as:

```
        gds_user = build(:user, permissions: [
          User::ACCESS_LIMIT_OVERRIDE_PERMISSION,
        ])
```

which then needs to be-rewritten as:

```
        gds_user = build(:user,
                         permissions: [
                           User::ACCESS_LIMIT_OVERRIDE_PERMISSION,
                         ])
```

which is less concise and less conventional with common GOV.UK patterns.