Disable Layout/MultilineMethodArgumentLineBreaks
Previously we enabled this Cop to bring more consistency to our
method calls, for which we have a variety of styles across our repos.
While array and hash items are often uniform in nature, this Cop had
undesirable outcomes for method calls with complex arguments [1].

The vertical overlap of arguments and array items makes them harder to
visually parse, compared to the original.

+      build(:document_type,
+            tags: [primary_organisation_field,
+                   organisations_field,
+                   role_appointments_field])

vs.

+      build(:document_type, tags: [primary_organisation_field,
                                    organisations_field,
                                    role_appointments_field])

[1]: https://github.com/alphagov/rubocop-govuk/pull/63#issuecomment-632002471