Lock dependencies to their full versions
Previously we only specified higher order versions for some of the
dependencies of this repo, which lead to varying versions being
used downstream, and a lack of DependaBot PRs because there's no
corresponding lockfile to record the version bump. This specifies
the full version of each dependent gem, except for dev dependencies.