Release version 4.1.0
This release contains updates of the various dependencies.

From my testing, the changes are relatively minor. Testing it on GOV.UK
repos, I've been able to apply this to the following without any
changes needed:

- smart-answers
- content-publisher
- email-alert-api
- search-api
- publisher

These repos have changes that can be resolved with `rubocop -A`:

- whitehall
- frontend

The common issue that I've seen in test cases is that Rubocop now
prefers `['Item A', 'Item B']` over `%W[Item\ A Item\ B]` - which seems
fair enough as the latter is a bit cryptic.